+++ src/middleware/language-switch.js
@@ -1,4 +1,6 @@
-export default async ({ app, store }) => {
-	const lang = store.getters["auth/getLocale"];
+import AuthModule from "@/store/auth";
+
+export default async ({ app }) => {
+	const lang = AuthModule.getLocale;
 	if (lang) app.i18n.locale = lang;
 };