+++ src/pages/administration/teachers/index.vue
@@ -128,7 +128,9 @@
 	</section>
 </template>
 <script>
+import AuthModule from "@/store/auth";
 import { mapGetters } from "vuex";
+import SchoolsModule from "@/store/schools";
 import EnvConfigModule from "@/store/env-config";
 import BackendDataTable from "@components/organisms/DataTable/BackendDataTable";
 import AdminTableLegend from "@components/molecules/AdminTableLegend";
@@ -274,12 +276,6 @@ export default {
 		};
 	},
 	computed: {
-		...mapGetters("auth", {
-			user: "getUser",
-		}),
-		...mapGetters("schools", {
-			schoolIsExternallyManaged: "schoolIsExternallyManaged",
-		}),
 		...mapGetters("users", {
 			teachers: "getList",
 			pagination: "getPagination",
@@ -287,6 +283,12 @@ export default {
 			deletedPercent: "getPercent",
 			qrLinks: "getQrLinks",
 		}),
+		user() {
+			return AuthModule.getUser;
+		},
+		schoolIsExternallyManaged() {
+			return SchoolsModule.schoolIsExternallyManaged;
+		},
 		env() {
 			return EnvConfigModule.getEnv;
 		},